Support Additional Claims in Service Credentials (#2)
To support authenticating requests to applications behind IAP, the
access token needs to encode an additional claim for the
"target_audience" representing the client ID of the IAP instance. This
also updates the version of Signet used for the OAuth client to a
version that supports signing JWTs with addiitional claims.