Add instructions for 3-legged-oauth
I spent a *long* time figuring out how to do this properly. The docs on
domain-wide delegation don't really make it clear that the `sub` has to
be set in order to take advantage of those settings. It seems to me like
this is the easiest way to run domain-wide delegation with this gem, but
if there's an easier way I'm all ears!