Fix: Don't raise an error if `expiration_time_millis` is nil; fall back to 0
`data.fetch("expiration_time_millis", 0)` will default to 0 only if the key is not present. If the key is present but the associated value is nil, then prior to this change the code would raise as follows:

```
NoMethodError (undefined method `/' for nil:NilClass

          expires_at:    data.fetch("expiration_time_millis", 0) / 1000
                                                                 ^):

```

This commit updates the code to default to 0 if the key is not found or if the value is nil.

With the existing token stores, this situation may not arise, but it's a particular issue when using a custom token store where values could be nil, like a database.